backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
icalendar==5.0.11
minio==7.2.0
authlib==1.3.0
httpx==0.25.2
python-dotenv==1.0.0
itsdangerous==2.1.2
